packages/flask-mono/requirements.txt
absl-py==2.2.2
astunparse==1.6.3
beautifulsoup4==4.13.4
blinker @ file:///croot/blinker_1737448726027/work
certifi==2025.1.31
charset-normalizer==3.4.1
click @ file:///croot/click_1744271578095/work
deepface==0.0.93
filelock==3.18.0
fire==0.7.0
Flask @ file:///croot/flask_1737454299809/work
flask-cors==5.0.1
flatbuffers==25.2.10
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==23.0.0
h5py==3.13.0
idna==3.10
itsdangerous @ file:///croot/itsdangerous_1716533337253/work
Jinja2 @ file:///croot/jinja2_1741710844255/work
joblib==1.4.2
keras==3.9.2
libclang==18.1.1
lz4==4.4.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
mdurl==0.1.2
ml_dtypes==0.5.1
mtcnn==1.0.0
namex==0.0.8
numpy==2.1.3
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
protobuf==5.29.4
Pygments==2.19.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
retina-face==0.0.17
rich==14.0.0
setuptools==75.8.0
six==1.17.0
soupsieve==2.7
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
tf_keras==2.19.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug @ file:///croot/werkzeug_1737448706492/work
wheel==0.45.1
wrapt==1.17.2
